+++ hrp/convert_test.go
@@ -4,25 +4,184 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+
+	"github.com/httprunner/httprunner/hrp/internal/builtin"
+)
+
+const (
+	templatesDir   = "internal/scaffold/templates/"
+	hrpExamplesDir = "../examples/hrp"
 )
 
 var (
-	demoTestCaseJSONPath    TestCasePath = "../examples/hrp/demo.json"
-	demoTestCaseYAMLPath    TestCasePath = "../examples/hrp/demo.yaml"
-	demoRefAPIYAMLPath      TestCasePath = "../examples/hrp/ref_api_test.yaml"
-	demoRefTestCaseJSONPath TestCasePath = "../examples/hrp/ref_testcase_test.json"
-	demoThinkTimeJsonPath   TestCasePath = "../examples/hrp/think_time_test.json"
-	demoAPIYAMLPath         APIPath      = "../examples/hrp/api/put.yml"
+	demoTestCaseWithPluginJSONPath    TestCasePath = templatesDir + "testcases/demo_with_funplugin.json"
+	demoTestCaseWithPluginYAMLPath    TestCasePath = templatesDir + "testcases/demo_with_funplugin.yaml"
+	demoTestCaseWithoutPluginJSONPath TestCasePath = templatesDir + "testcases/demo_without_funplugin.json"
+	demoTestCaseWithoutPluginYAMLPath TestCasePath = templatesDir + "testcases/demo_without_funplugin.yaml"
+	demoTestCaseWithRefAPIPath        TestCasePath = templatesDir + "testcases/demo_ref_api.json"
+	demoAPIGETPath                    APIPath      = templatesDir + "/api/get.yml"
 )
+
+var (
+	demoTestCaseWithThinkTimePath TestCasePath = hrpExamplesDir + "/think_time_test.json"
+)
+
+var demoTestCaseWithPlugin = &TestCase{
+	Config: NewConfig("demo with complex mechanisms").
+		SetBaseURL("https://postman-echo.com").
+		WithVariables(map[string]interface{}{ // global level variables
+			"n":       "${sum_ints(1, 2, 2)}",
+			"a":       "${sum(10, 2.3)}",
+			"b":       3.45,
+			"varFoo1": "${gen_random_string($n)}",
+			"varFoo2": "${max($a, $b)}", // 12.3; eval with built-in function
+		}),
+	TestSteps: []IStep{
+		NewStep("transaction 1 start").StartTransaction("tran1"), // start transaction
+		NewStep("get with params").
+			WithVariables(map[string]interface{}{ // step level variables
+				"n":       3,                // inherit config level variables if not set in step level, a/varFoo1
+				"b":       34.5,             // override config level variable if existed, n/b/varFoo2
+				"varFoo2": "${max($a, $b)}", // 34.5; override variable b and eval again
+				"name":    "get with params",
+			}).
+			SetupHook("${setup_hook_example($name)}").
+			GET("/get").
+			TeardownHook("${teardown_hook_example($name)}").
+			WithParams(map[string]interface{}{"foo1": "$varFoo1", "foo2": "$varFoo2"}). // request with params
+			WithHeaders(map[string]string{"User-Agent": "HttpRunnerPlus"}).             // request with headers
+			Extract().
+			WithJmesPath("body.args.foo1", "varFoo1"). // extract variable with jmespath
+			Validate().
+			AssertEqual("status_code", 200, "check response status code").        // validate response status code
+			AssertStartsWith("headers.\"Content-Type\"", "application/json", ""). // validate response header
+			AssertLengthEqual("body.args.foo1", 5, "check args foo1").            // validate response body with jmespath
+			AssertLengthEqual("$varFoo1", 5, "check args foo1").                  // assert with extracted variable from current step
+			AssertEqual("body.args.foo2", "34.5", "check args foo2"),             // notice: request params value will be converted to string
+		NewStep("transaction 1 end").EndTransaction("tran1"), // end transaction
+		NewStep("post json data").
+			POST("/post").
+			WithBody(map[string]interface{}{
+				"foo1": "$varFoo1",       // reference former extracted variable
+				"foo2": "${max($a, $b)}", // 12.3; step level variables are independent, variable b is 3.45 here
+			}).
+			Validate().
+			AssertEqual("status_code", 200, "check status code").
+			AssertLengthEqual("body.json.foo1", 5, "check args foo1").
+			AssertEqual("body.json.foo2", 12.3, "check args foo2"),
+		NewStep("post form data").
+			POST("/post").
+			WithHeaders(map[string]string{"Content-Type": "application/x-www-form-urlencoded; charset=UTF-8"}).
+			WithBody(map[string]interface{}{
+				"foo1": "$varFoo1",       // reference former extracted variable
+				"foo2": "${max($a, $b)}", // 12.3; step level variables are independent, variable b is 3.45 here
+				"time": "${get_timestamp()}",
+			}).
+			Extract().
+			WithJmesPath("body.form.time", "varTime").
+			Validate().
+			AssertEqual("status_code", 200, "check status code").
+			AssertLengthEqual("body.form.foo1", 5, "check args foo1").
+			AssertEqual("body.form.foo2", "12.3", "check args foo2"), // form data will be converted to string
+		NewStep("get with timestamp").
+			GET("/get").WithParams(map[string]interface{}{"time": "$varTime"}).
+			Validate().
+			AssertLengthEqual("body.args.time", 13, "check extracted var timestamp"),
+	},
+}
+
+var demoTestCaseWithoutPlugin = &TestCase{
+	Config: NewConfig("demo without custom function plugin").
+		SetBaseURL("https://postman-echo.com").
+		WithVariables(map[string]interface{}{ // global level variables
+			"n":       5,
+			"a":       12.3,
+			"b":       3.45,
+			"varFoo1": "${gen_random_string($n)}",
+			"varFoo2": "${max($a, $b)}", // 12.3; eval with built-in function
+		}),
+	TestSteps: []IStep{
+		NewStep("transaction 1 start").StartTransaction("tran1"), // start transaction
+		NewStep("get with params").
+			WithVariables(map[string]interface{}{ // step level variables
+				"n":       3,                // inherit config level variables if not set in step level, a/varFoo1
+				"b":       34.5,             // override config level variable if existed, n/b/varFoo2
+				"varFoo2": "${max($a, $b)}", // 34.5; override variable b and eval again
+				"name":    "get with params",
+			}).
+			GET("/get").
+			WithParams(map[string]interface{}{"foo1": "$varFoo1", "foo2": "$varFoo2"}). // request with params
+			WithHeaders(map[string]string{"User-Agent": "HttpRunnerPlus"}).             // request with headers
+			Extract().
+			WithJmesPath("body.args.foo1", "varFoo1"). // extract variable with jmespath
+			Validate().
+			AssertEqual("status_code", 200, "check response status code").        // validate response status code
+			AssertStartsWith("headers.\"Content-Type\"", "application/json", ""). // validate response header
+			AssertLengthEqual("body.args.foo1", 5, "check args foo1").            // validate response body with jmespath
+			AssertLengthEqual("$varFoo1", 5, "check args foo1").                  // assert with extracted variable from current step
+			AssertEqual("body.args.foo2", "34.5", "check args foo2"),             // notice: request params value will be converted to string
+		NewStep("transaction 1 end").EndTransaction("tran1"), // end transaction
+		NewStep("post json data").
+			POST("/post").
+			WithBody(map[string]interface{}{
+				"foo1": "$varFoo1",       // reference former extracted variable
+				"foo2": "${max($a, $b)}", // 12.3; step level variables are independent, variable b is 3.45 here
+			}).
+			Validate().
+			AssertEqual("status_code", 200, "check status code").
+			AssertLengthEqual("body.json.foo1", 5, "check args foo1").
+			AssertEqual("body.json.foo2", 12.3, "check args foo2"),
+		NewStep("post form data").
+			POST("/post").
+			WithHeaders(map[string]string{"Content-Type": "application/x-www-form-urlencoded; charset=UTF-8"}).
+			WithBody(map[string]interface{}{
+				"foo1": "$varFoo1",       // reference former extracted variable
+				"foo2": "${max($a, $b)}", // 12.3; step level variables are independent, variable b is 3.45 here
+				"time": "${get_timestamp()}",
+			}).
+			Extract().
+			WithJmesPath("body.form.time", "varTime").
+			Validate().
+			AssertEqual("status_code", 200, "check status code").
+			AssertLengthEqual("body.form.foo1", 5, "check args foo1").
+			AssertEqual("body.form.foo2", "12.3", "check args foo2"), // form data will be converted to string
+		NewStep("get with timestamp").
+			GET("/get").WithParams(map[string]interface{}{"time": "$varTime"}).
+			Validate().
+			AssertLengthEqual("body.args.time", 13, "check extracted var timestamp"),
+	},
+}
+
+func TestGenDemoTestCase(t *testing.T) {
+	tCase, _ := demoTestCaseWithPlugin.ToTCase()
+	err := builtin.Dump2JSON(tCase, demoTestCaseWithPluginJSONPath.GetPath())
+	if err != nil {
+		t.Fail()
+	}
+	err = builtin.Dump2YAML(tCase, demoTestCaseWithPluginYAMLPath.GetPath())
+	if err != nil {
+		t.Fail()
+	}
+
+	tCase, _ = demoTestCaseWithoutPlugin.ToTCase()
+	err = builtin.Dump2JSON(tCase, demoTestCaseWithoutPluginJSONPath.GetPath())
+	if err != nil {
+		t.Fail()
+	}
+	err = builtin.Dump2YAML(tCase, demoTestCaseWithoutPluginYAMLPath.GetPath())
+	if err != nil {
+		t.Fail()
+	}
+}
 
 func TestLoadCase(t *testing.T) {
 	tcJSON := &TCase{}
 	tcYAML := &TCase{}
-	err := loadFromJSON(demoTestCaseJSONPath.ToString(), tcJSON)
+	err := builtin.LoadFile(demoTestCaseWithPluginJSONPath.GetPath(), tcJSON)
 	if !assert.NoError(t, err) {
 		t.Fail()
 	}
-	err = loadFromYAML(demoTestCaseYAMLPath.ToString(), tcYAML)
+	err = builtin.LoadFile(demoTestCaseWithPluginYAMLPath.GetPath(), tcYAML)
 	if !assert.NoError(t, err) {
 		t.Fail()
 	}
@@ -41,7 +200,7 @@ func TestLoadCase(t *testing.T) {
 	}
 }
 
-func Test_convertCheckExpr(t *testing.T) {
+func TestConvertCheckExpr(t *testing.T) {
 	exprs := []struct {
 		before string
 		after  string